pyproject.toml
[project]
name = "reddit-np-topic-modeling"
version = "0.1.0"
description = "Multiscale topic modeling of Reddit discussions on US National Parks"
requires-python = ">=3.10"

dependencies = [
    "bertopic>=0.16",
    "sentence-transformers>=2.7",
    "gliner>=0.2",
    "duckdb>=0.10",
    "umap-learn>=0.5",
    "hdbscan>=0.8",
    "swifter>=1.4",
    "pandas>=2.0",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "nltk>=3.8",
    "plotly>=5.0",
    "pyyaml>=6.0",
    "jupyterlab>=4.0",
    "ipywidgets>=8.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reddit_np_topics"]